+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/authentication/ChangePassword.kt
@@ -4,9 +4,10 @@ import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import net.folivo.trixnity.clientserverapi.model.uia.MatrixUIAEndpoint
+import net.folivo.trixnity.core.Auth
+import net.folivo.trixnity.core.AuthRequired
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.POST
-import net.folivo.trixnity.core.WithoutAuth
 
 /**
  * @see <a href="https://spec.matrix.org/v1.10/client-server-api/#post_matrixclientv3accountpassword">matrix spec</a>
@@ -14,7 +15,7 @@ import net.folivo.trixnity.core.WithoutAuth
 @Serializable
 @Resource("/_matrix/client/v3/account/password")
 @HttpMethod(POST)
-@WithoutAuth(true)
+@Auth(AuthRequired.OPTIONAL)
 object ChangePassword : MatrixUIAEndpoint<ChangePassword.Request, Unit> {
     @Serializable
     data class Request(