+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/media/DownloadMediaLegacy.kt
@@ -5,10 +5,11 @@ import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
+import net.folivo.trixnity.core.Auth
+import net.folivo.trixnity.core.AuthRequired
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.GET
 import net.folivo.trixnity.core.MatrixEndpoint
-import net.folivo.trixnity.core.WithoutAuth
 
 /**
  * @see <a href="https://spec.matrix.org/v1.10/client-server-api/#get_matrixmediav3downloadservernamemediaid">matrix spec</a>
@@ -16,7 +17,7 @@ import net.folivo.trixnity.core.WithoutAuth
 @Serializable
 @Resource("/_matrix/media/v3/download/{serverName}/{mediaId}")
 @HttpMethod(GET)
-@WithoutAuth
+@Auth(AuthRequired.NO)
 @Deprecated("use DownloadMedia instead")
 data class DownloadMediaLegacy(
     @SerialName("serverName") val serverName: String,