+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/server/GetVersions.kt
@@ -3,10 +3,11 @@ package net.folivo.trixnity.clientserverapi.model.server
 import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import net.folivo.trixnity.core.Auth
+import net.folivo.trixnity.core.AuthRequired
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.GET
 import net.folivo.trixnity.core.MatrixEndpoint
-import net.folivo.trixnity.core.WithoutAuth
 
 /**
  * @see <a href="https://spec.matrix.org/v1.10/client-server-api/#get_matrixclientversions">matrix spec</a>
@@ -14,7 +15,7 @@ import net.folivo.trixnity.core.WithoutAuth
 @Serializable
 @Resource("/_matrix/client/versions")
 @HttpMethod(GET)
-@WithoutAuth(true)
+@Auth(AuthRequired.OPTIONAL)
 object GetVersions : MatrixEndpoint<Unit, GetVersions.Response> {
     @Serializable
     data class Response(