+++ trixnity-serverserverapi/trixnity-serverserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/serverserverapi/model/discovery/GetServerVersion.kt
@@ -3,10 +3,11 @@ package net.folivo.trixnity.serverserverapi.model.discovery
 import io.ktor.resources.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import net.folivo.trixnity.core.Auth
+import net.folivo.trixnity.core.AuthRequired
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.GET
 import net.folivo.trixnity.core.MatrixEndpoint
-import net.folivo.trixnity.core.WithoutAuth
 
 /**
  * @see <a href="https://spec.matrix.org/v1.10/server-server-api/#get_matrixfederationv1version">matrix spec</a>
@@ -14,7 +15,7 @@ import net.folivo.trixnity.core.WithoutAuth
 @Serializable
 @Resource("/_matrix/federation/v1/version")
 @HttpMethod(GET)
-@WithoutAuth
+@Auth(AuthRequired.NO)
 object GetServerVersion : MatrixEndpoint<Unit, GetServerVersion.Response> {
     @Serializable
     data class Response(