+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Reset unread marker on sending own message.
+- Support for refresh tokens.
 
 ### Changed